refactor(web): extract API client and auth flow into dedicated modules
Move the real API implementation from client.ts into real-api.ts and
extract the login/session/logout flow from auth.ts into api-auth-flow.ts.
Route direct api imports through getApiClient dispatcher for consistency.

- real-api.ts: HTTP fetch logic, request builders, response parsing
- api-auth-flow.ts: credential login, session verification, logout
- client.ts: types, dispatcher, re-exports
- auth.ts: shared state management, delegates to api-auth-flow
- Update hooks to accept apiClient parameter instead of importing directly
